aggregates tickets #325 + #341 and TSN main-ticket #342: we committed to use distinct IP addresses inside a consist, even over different VLANs and we use preconfigured VLANs and PCP priorities to select preconfigured TSN timeslots on the network
git-svn-id: https://svn.code.sf.net/p/tcnopen/trdp/trunk@2384 3b5a3598-5f4e-4449-9e63-bd40438bfec0

+++ trdp/src/common/trdp_pdindex.c
@@ -1214,7 +1214,7 @@ TRDP_ERR_T trdp_pdSendIndexed (TRDP_SESSION_PT appHandle)
             timeDiff = clockTimeStamp;
             vos_subTime(&timeDiff, &pSlot->latestCycleStartTimeStamp);
             clockTimeSpentInCycle = timeDiff.tv_sec * 1000000 + timeDiff.tv_usec;
-            percentClockUsed = 100.0 * clockTimeSpentInCycle / pSlot->currentCycle;  /* optimal: 100%, on slow hardware > 100% */
+            percentClockUsed = (REAL32)100.0 * (REAL32)(clockTimeSpentInCycle / pSlot->currentCycle);  /* optimal: 100%, on slow hardware > 100% */
 
             if (percentClockUsed > CLOCK_PERCENT_ERROR_LIMIT) {            /* consider to improve setup */
                 logLevel = VOS_LOG_ERROR;
